+++ pymc3/tests/test_math.py
@@ -1,9 +1,11 @@
 import numpy as np
+import numpy.testing as npt
 import theano
 import theano.tensor as tt
 from theano.tests import unittest_tools as utt
 from pymc3.math import (
-    LogDet, logdet, probit, invprobit, expand_packed_triangular)
+    LogDet, logdet, probit, invprobit, expand_packed_triangular,
+    log1pexp, log1mexp)
 from .helpers import SeededTest
 import pytest
 from pymc3.theanof import floatX
@@ -12,6 +14,42 @@ from pymc3.theanof import floatX
 def test_probit():
     p = np.array([0.01, 0.25, 0.5, 0.75, 0.99])
     np.testing.assert_allclose(invprobit(probit(p)).eval(), p, atol=1e-5)
+
+
+def test_log1pexp():
+    vals = np.array([-1e20, -100, -10, -1e-4, 0, 1e-4, 10, 100, 1e20])
+    # import mpmath
+    # mpmath.mp.dps = 1000
+    # [float(mpmath.log(1 + mpmath.exp(x))) for x in vals]
+    expected = np.array([
+        0.0,
+        3.720075976020836e-44,
+        4.539889921686465e-05,
+        0.6930971818099453,
+        0.6931471805599453,
+        0.6931971818099453,
+        10.000045398899218,
+        100.0,
+        1e+20])
+    actual = log1pexp(vals).eval()
+    npt.assert_allclose(actual, expected)
+
+
+def test_log1mexp():
+    vals = np.array([-1, 0, 1e-20, 1e-4, 10, 100, 1e20])
+    # import mpmath
+    # mpmath.mp.dps = 1000
+    # [float(mpmath.log(1 - mpmath.exp(-x))) for x in vals]
+    expected = np.array([
+        np.nan,
+        -np.inf,
+        -46.051701859880914,
+        -9.210390371559516,
+        -4.540096037048921e-05,
+        -3.720075976020836e-44,
+        0.0])
+    actual = log1mexp(vals).eval()
+    npt.assert_allclose(actual, expected)
 
 
 class TestLogDet(SeededTest):